# openapi.yaml
openapi: 3.1.0
info:
  title: Insight Store — Telegram Forwarding Bot
  description: Automatic Telegram forwarding between authorized chats. No quotas,
    no subscription limits — and no evasion of Telegram's own limits.
  version: 0.1.0
paths:
...
  /api/v1/forwarding-rules:
    get:
      tags:
      - rules
      summary: List Rules
      operationId: list_rules_api_v1_forwarding_rules_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RuleListItem'
                type: array
                title: Response List Rules Api V1 Forwarding Rules Get
    post:
      tags:
      - rules
      summary: Create Rule
      operationId: create_rule_api_v1_forwarding_rules_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleWriteRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/forwarding-rules/{rule_id}:
    get:
      tags:
      - rules
      summary: Get Rule
      operationId: get_rule_api_v1_forwarding_rules__rule_id__get
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rule Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - rules
      summary: Update Rule
      operationId: update_rule_api_v1_forwarding_rules__rule_id__patch
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rule Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleWriteRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - rules
      summary: Delete Rule
      operationId: delete_rule_api_v1_forwarding_rules__rule_id__delete
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rule Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/forwarding-rules/{rule_id}/activate:
    post:
      tags:
      - rules
      summary: Activate Rule
      operationId: activate_rule_api_v1_forwarding_rules__rule_id__activate_post
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rule Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/forwarding-rules/{rule_id}/pause:
    post:
      tags:
      - rules
      summary: Pause Rule
      operationId: pause_rule_api_v1_forwarding_rules__rule_id__pause_post
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rule Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/forwarding-rules/{rule_id}/resume:
    post:
      tags:
      - rules
      summary: Resume Rule
      operationId: resume_rule_api_v1_forwarding_rules__rule_id__resume_post
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rule Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/forwarding-rules/{rule_id}/events:
    get:
      tags:
      - rules
      summary: List Rule Events
      operationId: list_rule_events_api_v1_forwarding_rules__rule_id__events_get
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rule Id
      - name: outcome
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Outcome
      - name: destination_chat_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Destination Chat Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponse'
                title: Response List Rule Events Api V1 Forwarding Rules  Rule Id  Events
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/forwarding-rules/{rule_id}/retry-failed:
    post:
      tags:
      - rules
      summary: Retry Failed
      description: Requeues only failed / needs-attention work. Successes are never
        replayed.
      operationId: retry_failed_api_v1_forwarding_rules__rule_id__retry_failed_post
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rule Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/forwarding-rules/{rule_id}/jobs:
    get:
      tags:
      - rules
      summary: List Rule Jobs
      description: Per-destination status, which is what the Rule Detail page renders.
      operationId: list_rule_jobs_api_v1_forwarding_rules__rule_id__jobs_get
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rule Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Rule Jobs Api V1 Forwarding Rules  Rule Id  Jobs
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AcceptedResponse:
      properties:
        status:
          type: string
          const: accepted
          title: Status
          default: accepted
        task_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Task Id
        message:
          type: string
          title: Message
          default: Queued. Refresh to see the result.
      type: object
      title: AcceptedResponse
      description: 202 body for anything that enqueues background work.
...
    EventResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        rule_id:
          type: string
          format: uuid
          title: Rule Id
        outcome:
          type: string
          title: Outcome
        reason_code:
          type: string
          title: Reason Code
        reason_text:
          type: string
          title: Reason Text
          default: ''
        detail_safe:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail Safe
        attempt:
          type: integer
          title: Attempt
        source_chat_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Source Chat Id
        destination_chat_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Destination Chat Id
        source_message_ids:
          items:
            type: string
          type: array
          title: Source Message Ids
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
      type: object
      required:
      - id
      - rule_id
      - outcome
      - reason_code
      - detail_safe
      - attempt
      - source_chat_id
      - destination_chat_id
      - occurred_at
      title: EventResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RuleChatSummary:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        peer_id:
          type: string
          title: Peer Id
        peer_type:
          type: string
          title: Peer Type
        eligible:
          type: boolean
          title: Eligible
        reason_code:
          type: string
          title: Reason Code
      type: object
      required:
      - id
      - title
      - peer_id
      - peer_type
      - eligible
      - reason_code
      title: RuleChatSummary
    RuleListItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        status:
          type: string
          title: Status
        connection_id:
          type: string
          format: uuid
          title: Connection Id
        destination_count:
          type: integer
          title: Destination Count
        source_titles:
          items:
            type: string
          type: array
          title: Source Titles
        filter_summary:
          type: string
          title: Filter Summary
        last_activity_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Activity At
        paused_reason_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Paused Reason Text
      type: object
      required:
      - id
      - name
      - status
      - connection_id
      - destination_count
      - source_titles
      - filter_summary
      - last_activity_at
      title: RuleListItem
    RuleResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        connection_id:
          type: string
          format: uuid
          title: Connection Id
        status:
          type: string
          title: Status
        version:
          type: integer
          title: Version
        forward_mode:
          type: string
          title: Forward Mode
        delay_ms:
          type: integer
          title: Delay Ms
        keyword_include:
          items:
            type: string
          type: array
          title: Keyword Include
        keyword_exclude:
          items:
            type: string
          type: array
          title: Keyword Exclude
        keyword_match_mode:
          type: string
          title: Keyword Match Mode
        media_types:
          items:
            type: string
          type: array
          title: Media Types
        preserve_links:
          type: boolean
          title: Preserve Links
        preserve_caption:
          type: boolean
          title: Preserve Caption
        paused_reason_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Paused Reason Code
        paused_reason_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Paused Reason Text
        last_activity_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Activity At
        created_at:
          type: string
          format: date-time
          title: Created At
        sources:
          items:
            $ref: '#/components/schemas/RuleChatSummary'
          type: array
          title: Sources
        destinations:
          items:
            $ref: '#/components/schemas/RuleChatSummary'
          type: array
          title: Destinations
        preview:
          type: string
          title: Preview
          default: ''
      type: object
      required:
      - id
      - name
      - connection_id
      - status
      - version
      - forward_mode
      - delay_ms
      - keyword_include
      - keyword_exclude
      - keyword_match_mode
      - media_types
      - preserve_links
      - preserve_caption
      - paused_reason_code
      - last_activity_at
      - created_at
      title: RuleResponse
    RuleWriteRequest:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        connection_id:
          type: string
          format: uuid
          title: Connection Id
        source_chat_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 50
          minItems: 1
          title: Source Chat Ids
        destination_chat_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 200
          minItems: 1
          title: Destination Chat Ids
        forward_mode:
          type: string
          enum:
          - forward
          - copy
          title: Forward Mode
          default: forward
        delay_ms:
          type: integer
          maximum: 3600000.0
          minimum: 0.0
          title: Delay Ms
          default: 0
        keyword_include:
          items:
            type: string
            maxLength: 128
          type: array
          title: Keyword Include
        keyword_exclude:
          items:
            type: string
            maxLength: 128
          type: array
          title: Keyword Exclude
        keyword_match_mode:
          type: string
          enum:
          - substring
          - word
          title: Keyword Match Mode
          default: substring
        media_types:
          items:
            type: string
          type: array
          title: Media Types
        preserve_links:
          type: boolean
          title: Preserve Links
          default: true
        preserve_caption:
          type: boolean
          title: Preserve Caption
          default: true
        allow_source_as_destination:
          type: boolean
          title: Allow Source As Destination
          default: false
      type: object
      required:
      - name
      - connection_id
      - source_chat_ids
      - destination_chat_ids
      title: RuleWriteRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...